Moved Dialogs to components folder and made dropdown more generic
