+++ app/src/main/java/org/listenbrainz/android/ui/components/dialogs/DialogText.kt
@@ -1,4 +1,4 @@
-package org.listenbrainz.android.ui.screens.feed.dialogs
+package org.listenbrainz.android.ui.components.dialogs
 
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
