+++ app/src/main/java/org/listenbrainz/android/ui/screens/feed/events/PersonalRecommendationFeedLayout.kt
@@ -20,9 +20,9 @@ import org.listenbrainz.android.model.Metadata
 import org.listenbrainz.android.model.feed.FeedEvent
 import org.listenbrainz.android.model.feed.FeedEventType
 import org.listenbrainz.android.ui.components.ListenCardSmall
+import org.listenbrainz.android.ui.components.dialogs.UserTag
 import org.listenbrainz.android.ui.screens.feed.BaseFeedLayout
-import org.listenbrainz.android.ui.screens.feed.FeedSocialDropdown
-import org.listenbrainz.android.ui.screens.feed.dialogs.UserTag
+import org.listenbrainz.android.ui.screens.feed.SocialDropdown
 import org.listenbrainz.android.ui.theme.ListenBrainzTheme
 import org.listenbrainz.android.util.Utils
 
@@ -59,9 +59,9 @@ fun PersonalRecommendationFeedLayout(
             enableDropdownIcon = true,
             onDropdownIconClick = onDropdownClick,
             dropDown = {
-                FeedSocialDropdown(
+                SocialDropdown(
                     isExpanded = dropdownState == index,
-                    event = event,
+                    metadata = event.metadata,
                     onDismiss = onDropdownClick,
                     onOpenInMusicBrainz = onOpenInMusicBrainz,
                     onPin = onPin,
