+++ app/src/main/java/org/listenbrainz/android/ui/screens/feed/events/PinFeedLayout.kt
@@ -12,7 +12,7 @@ import org.listenbrainz.android.model.feed.FeedEvent
 import org.listenbrainz.android.model.feed.FeedEventType
 import org.listenbrainz.android.ui.components.ListenCardSmall
 import org.listenbrainz.android.ui.screens.feed.BaseFeedLayout
-import org.listenbrainz.android.ui.screens.feed.FeedSocialDropdown
+import org.listenbrainz.android.ui.screens.feed.SocialDropdown
 import org.listenbrainz.android.ui.theme.ListenBrainzTheme
 import org.listenbrainz.android.util.Utils
 
@@ -52,10 +52,10 @@ fun PinFeedLayout(
             enableDropdownIcon = true,
             onDropdownIconClick = onDropdownClick,
             dropDown = {
-                FeedSocialDropdown(
+                SocialDropdown(
                     isExpanded = dropdownState == index,
                     onDismiss = onDropdownClick,
-                    event = event,
+                    metadata = event.metadata,
                     onOpenInMusicBrainz = onOpenInMusicBrainz,
                     onPin = onPin,
                     onRecommend = onRecommend,
